core: file reorg
