+++ core/src/main/kotlin/ch/kleis/lcaac/core/datasource/DefaultDataSourceOperations.kt
@@ -2,6 +2,10 @@ package ch.kleis.lcaac.core.datasource
 
 import ch.kleis.lcaac.core.config.DataSourceConfig
 import ch.kleis.lcaac.core.config.LcaacConfig
+import ch.kleis.lcaac.core.datasource.cache.ConnectorCache
+import ch.kleis.lcaac.core.datasource.cache.GetAllRequest
+import ch.kleis.lcaac.core.datasource.cache.GetFirstRequest
+import ch.kleis.lcaac.core.datasource.cache.SumProductRequest
 import ch.kleis.lcaac.core.datasource.in_memory.InMemoryConnector
 import ch.kleis.lcaac.core.lang.evaluator.reducer.DataExpressionReducer
 import ch.kleis.lcaac.core.lang.expression.*
@@ -12,56 +16,16 @@ import ch.kleis.lcaac.core.lang.value.RecordValue
 import ch.kleis.lcaac.core.lang.value.StringValue
 import ch.kleis.lcaac.core.math.QuantityOperations
 import ch.kleis.lcaac.core.prelude.Prelude
-import com.mayakapps.kache.InMemoryKache
-import com.mayakapps.kache.KacheStrategy
 import kotlinx.coroutines.runBlocking
 
 typealias ConnectorName = String
 typealias SourceName = String
 
-data class GetAllRequest<Q>(
-    val config: DataSourceConfig,
-    val source: DataSourceValue<Q>
-)
-
-data class GetFirstRequest<Q>(
-    val config: DataSourceConfig,
-    val source: DataSourceValue<Q>
-)
-
-data class SumProductRequest<Q>(
-    val config: DataSourceConfig,
-    val source: DataSourceValue<Q>,
-    val columns: List<String>,
-)
-
-
-class ConnectorCache<Q>(
-    private val maxSize: Long
-) {
-    val getAllCache = InMemoryKache<GetAllRequest<Q>, Sequence<ERecord<Q>>>(
-        maxSize = maxSize
-    ) {
-        strategy = KacheStrategy.LRU
-    }
-
-    val getFirstCache = InMemoryKache<GetFirstRequest<Q>, ERecord<Q>>(
-        maxSize = maxSize,
-    ) {
-        strategy = KacheStrategy.LRU
-    }
-    val sumProductCache = InMemoryKache<SumProductRequest<Q>, DataExpression<Q>>(
-        maxSize = maxSize,
-    ) {
-        strategy = KacheStrategy.LRU
-    }
-}
-
 class DefaultDataSourceOperations<Q>(
     private val ops: QuantityOperations<Q>,
     private val config: LcaacConfig,
     private val connectors: Map<ConnectorName, DataSourceConnector<Q>>,
-    private val overrides: Map<SourceName, ConnectorName>,
+    private val overrides: Map<SourceName, ConnectorName> = emptyMap(),
     private val cache: Map<String, ConnectorCache<Q>> = connectors
         .filter { it.value.getConfig().cache.enabled }
         .mapValues {
